Moved the validation method to intellisense operations
