+++ src/libraries/Microsoft.PowerFx.Core/Public/CheckResult.cs
@@ -6,9 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.PowerFx.Core.Binding;
 using Microsoft.PowerFx.Core.Errors;
-using Microsoft.PowerFx.Core.Parser;
 using Microsoft.PowerFx.Core.Public;
-using Microsoft.PowerFx.Core.Types;
 using Microsoft.PowerFx.Core.Utils;
 using Microsoft.PowerFx.Intellisense;
 using Microsoft.PowerFx.Syntax;
@@ -95,63 +93,6 @@ namespace Microsoft.PowerFx
         /// <param name="node"></param>
         /// <returns></returns>
         public FormulaType GetNodeType(TexlNode node) => FormulaType.Build(_binding.GetType(node));
-
-        /// <summary>
-        /// Checks whether a call to a function with name <paramref name="functionName" /> is valid with argument list
-        /// <paramref name="args" />. Additionally returns (as an out parameter) the return type of this invocation.
-        /// 
-        /// Note: all arguments must belong to the formula that belongs to this <see cref="CheckResult" />.
-        /// </summary>
-        /// <param name="functionName"></param>
-        /// <param name="args"></param>
-        /// <param name="retType"></param>
-        /// <returns></returns>
-        public bool ValidateInvocation(string functionName, IReadOnlyList<TexlNode> args, out FormulaType retType)
-        {
-            retType = null;
-
-            if (args == null)
-            {
-                throw new ArgumentNullException(nameof(args));
-            }
-
-            foreach ((var arg, var index) in args.Select((value, index) => (value, index)))
-            {
-                if (arg == null)
-                {
-                    throw new ArgumentNullException(nameof(args), $"Argument {index} is null");
-                }
-
-                if (!_binding.IsNodeValid(arg))
-                {
-                    throw new ArgumentException($"Argument {index} does not belong to this result");
-                }
-            }
-
-            var types = args.Select(node => _binding.GetType(node)).ToArray();
-
-            // TODO: Horrible hack to get function identifier name - how to do this in idiomatic way?
-            var fncIdent = TexlParser.ParseScript($"{functionName}()").Root.AsCall().Head;
-
-            // Note: there could be multiple functions (e.g., overloads) with the same name and arity,
-            //  hence loop through candidates and check whether one of them matches.
-            var fncs = _binding.NameResolver.Functions
-                              .Where(fnc => fnc.Name == fncIdent.Name && fnc.Namespace == fncIdent.Namespace
-                                                && args.Count >= fnc.MinArity && args.Count <= fnc.MaxArity);
-            foreach (var fnc in fncs)
-            {
-                var result =
-                    fnc.CheckInvocation(_binding, args.ToArray(), types, _binding.ErrorContainer, out var retDType, out _);
-
-                if (result)
-                {
-                    retType = FormulaType.Build(retDType);
-                    return true;
-                }
-            }
-
-            return false;
-        }
 
         internal IReadOnlyDictionary<string, TokenResultType> GetTokens(GetTokensFlags flags) => GetTokensUtils.GetTokens(_binding, flags);
     }
