+++ src/libraries/Microsoft.PowerFx.Core/Public/IntellisenseOperations.cs
@@ -3,7 +3,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
+using Microsoft.PowerFx.Core.Parser;
+using Microsoft.PowerFx.Syntax;
+using Microsoft.PowerFx.Types;
 
 namespace Microsoft.PowerFx.Intellisense
 {
@@ -21,6 +24,69 @@ namespace Microsoft.PowerFx.Intellisense
         public IntellisenseOperations(CheckResult result)
         {
             _checkResult = result;
+        }
+
+        /// <summary>
+        /// Checks whether a call to a function with name <paramref name="functionName" /> is valid with argument list
+        /// <paramref name="args" />. Additionally returns (as an out parameter) the return type of this invocation.
+        /// 
+        /// Note: all arguments must belong to the formula that belongs to this <see cref="CheckResult" />.
+        /// </summary>
+        /// <param name="functionName"></param>
+        /// <param name="args"></param>
+        /// <param name="retType"></param>
+        /// <returns></returns>
+        public bool ValidateInvocation(string functionName, IReadOnlyList<TexlNode> args, out FormulaType retType)
+        {
+            retType = null;
+
+            if (args == null)
+            {
+                throw new ArgumentNullException(nameof(args));
+            }
+
+            foreach ((var arg, var index) in args.Select((value, index) => (value, index)))
+            {
+                if (arg == null)
+                {
+                    throw new ArgumentNullException(nameof(args), $"Argument {index} is null");
+                }
+
+                if (!_checkResult._binding.IsNodeValid(arg))
+                {
+                    throw new ArgumentException($"Argument {index} does not belong to this result");
+                }
+            }
+
+            var types = args.Select(node => _checkResult._binding.GetType(node)).ToArray();
+
+            // TODO: Horrible hack to get function identifier name - how to do this in idiomatic way?
+            var fncIdent = TexlParser.ParseScript($"{functionName}()").Root.AsCall().Head;
+
+            // Note: there could be multiple functions (e.g., overloads) with the same name and arity,
+            //  hence loop through candidates and check whether one of them matches.
+            var fncs = _checkResult._binding.NameResolver.Functions
+                                   .Where(fnc => fnc.Name == fncIdent.Name && fnc.Namespace == fncIdent.Namespace
+                                                    && args.Count >= fnc.MinArity && args.Count <= fnc.MaxArity);
+            foreach (var fnc in fncs)
+            {
+                var result =
+                    fnc.CheckInvocation(
+                        _checkResult._binding,
+                        args.ToArray(),
+                        types,
+                        _checkResult._binding.ErrorContainer,
+                        out var retDType,
+                        out _);
+
+                if (result)
+                {
+                    retType = FormulaType.Build(retDType);
+                    return true;
+                }
+            }
+
+            return false;
         }
     }
 }
